refactor: add logger decoror in factories and standardize imports
