+++ src/main/factories/middlewares/auth-middleware-factory.ts
@@ -1,9 +1,9 @@
-import { AuthMiddleware } from '@/application/middlewares/auth'
+import { type Middleware } from '@/application/protocols'
+import { AuthMiddleware } from '@/application/middlewares'
 import { JwtAdapter } from '@/infra/cryptography/jwt-adapter'
 import { GuardianAccountRepository } from '@/infra/repos/postgresql/guardian-account-repository'
-import { type Middleware } from '@/application/middlewares/middleware'
+import { BcryptAdapter } from '@/infra/cryptography'
 import env from '@/main/config/env'
-import { BcryptAdapter } from '@/infra/cryptography/bcrypt-adapter'
 
 export const makeAuthMiddleware = (): Middleware => {
   const secret = env.secret
